Jest tests
Split animal reducer into new test suite.  Update tests.
